infra: move mersenne factory functions to util
