+++ src/index.ts
@@ -30,10 +30,6 @@ export type {
 export { FakerError } from './errors/faker-error';
 export { Faker } from './faker';
 export type { FakerOptions } from './faker';
-export {
-  generateMersenne32Randomizer,
-  generateMersenne53Randomizer,
-} from './internal/mersenne';
 export * from './locale';
 export { fakerEN as faker } from './locale';
 export * from './locales';
@@ -85,3 +81,7 @@ export type { WordModule } from './modules/word';
 export type { Randomizer } from './randomizer';
 export { SimpleFaker, simpleFaker } from './simple-faker';
 export { mergeLocales } from './utils/merge-locales';
+export {
+  generateMersenne32Randomizer,
+  generateMersenne53Randomizer,
+} from './utils/mersenne';
